remove directory and rename files
